fix: perbaikan unknown column type "tinyinteger" requested (#4536)

+++ donjo-app/models/migrations/Migrasi_2025020551.php
@@ -91,7 +91,7 @@ class Migrasi_2025020551
         DB::table('program')->whereNull('sasaran')->update(['sasaran' => 0]);
 
         Schema::table('program', static function (Blueprint $table) {
-            $table->tinyInteger('sasaran')->nullable(false)->change();
+            $table->integer('sasaran')->nullable(false)->change();
         });
     }
 }
